# doc/swagger.yml
openapi: 3.0.1
info:
  title: dating server API
  description: Описание методов API сервера мобильного приложения
  version: 0.24.6
servers:
  - url: https://dev.dating.eagledev.ru/api/v1
  - url: https://prod.dating.eagledev.ru/api/v1
  - url: http://localhost:8032/api/v1/
paths:
...
  /profile/my:
    get:
      tags:
        - Профиль пользователя
      summary: Мой профиль
      security:
        - bearerAuth: [ ]
      description: Получение данных своего профиля
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                required:
                  - data
                  - success
                type: object
                properties:
                  success:
                    type: boolean
                    description: статус запроса
                    example: true
                  data:
                    $ref: '#/components/schemas/UserProfile'
        "403":
          description: Ошибка доступа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: server error
          content: { }
    post:
      tags:
        - Профиль пользователя
      summary: Обновить профиль
      security:
        - bearerAuth: [ ]
      description: Обновление данных профиля пользователя
      requestBody:
        content:
          application/json:
            schema:
              required:
                - birthday
                - name
                - sex
              type: object
              properties:
                name:
                  type: string
                  description: Имя
                  example: Боб
                sex:
                  type: string
                  description: enum пол пользователя (male\female\undefined)
                  example: male
                birthday:
                  type: string
                  description: дата рождения в формате 2006-02-01
                  example: 1995-10-20
                city:
                  type: string
                  description: Город
                  example: Кисловодск
                about:
                  type: string
                  description: О себе
                  example: ем только куриный суп
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                required:
                  - success
                type: object
                properties:
                  success:
                    type: boolean
                    description: статус запроса
                    example: true
        "400":
          description: ошибка в теле запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Ошибка доступа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: server error
          content: { }
      x-codegen-request-body-name: data
    delete:
      tags:
        - Профиль пользователя
      summary: Удалить аккаунт
      security:
        - bearerAuth: [ ]
      description: Удаление своего акканута
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                required:
                  - success
                type: object
                properties:
                  success:
                    type: boolean
                    description: успех запроса
                    example: true
        "500":
          description: server error
          content: { }
  /profile/my/photos:
    get:
      tags:
        - Профиль пользователя
        - Фото
      summary: Фотографии пользователя
      security:
        - bearerAuth: [ ]
      description: Получение списка своих фото
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                required:
                  - data
                  - success
                type: object
                properties:
                  success:
                    type: boolean
                    description: успех запроса
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserPhoto'
        "500":
          description: server error
          content: { }
    post:
      tags:
        - Профиль пользователя
        - Фото
      summary: Загрузить фото
      security:
        - bearerAuth: [ ]
      description: Загрузка нового фото пользователя
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photo:
                  type: string
                  format: binary
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                required:
                  - data
                  - success
                type: object
                properties:
                  success:
                    type: boolean
                    description: успех запроса
                    example: true
                  data:
                    required:
                      - ID
                      - URL
                    type: object
                    properties:
                      ID:
                        type: string
                        description: uuid загруженной фотографии
                        example: string
                      URL:
                        type: string
                        description: ссылка на фото
                        example: path/pic.jpg
        "400":
          description: ошибка в теле запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: ошибка доступа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "413":
          description: Файл слишком большой
          content: { }
        "500":
          description: server error
          content: { }
  /profile/my/photos/{photoId}:
    delete:
      tags:
        - Профиль пользователя
        - Фото
      summary: Удалить фото
      security:
        - bearerAuth: [ ]
      description: Удаление фото пользователя по id фотографии
      parameters:
        - name: photoId
          in: path
          description: UUID фотографии
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                required:
                  - success
                type: object
                properties:
                  success:
                    type: boolean
                    description: успех запроса
                    example: true
        "500":
          description: server error
          content: { }
  /profile/my/photos/{photo_id}/set-main:
    get:
      tags:
        - Профиль пользователя
        - Фото
      summary: Установить фото как основное
      security:
        - bearerAuth: [ ]
      description: устанавливает данную фотографию как основную (аватарка)
      parameters:
        - name: photoId
          in: path
          description: UUID фотографии
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                required:
                  - success
                type: object
                properties:
                  success:
                    type: boolean
                    description: успех запроса
                    example: true
        "500":
          description: server error
          content: { }
  /profile/my/geo:
    post:
      tags:
        - Профиль пользователя
      summary: Обновить геолокацию
      security:
        - bearerAuth: [ ]
      description: Обновление текущей геолокации пользователя на сервере
      requestBody:
        content:
          application/json:
            schema:
              required:
                - lat
                - long
              type: object
              properties:
                lat:
                  type: number
                  description: latitude
                  example: 13.361389
                long:
                  type: number
                  description: longitude
                  example: 38.115556
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                required:
                  - success
                type: object
                properties:
                  success:
                    type: boolean
                    description: успех запроса
                    example: true
        "500":
          description: server error
          content: { }
      x-codegen-request-body-name: data
  /profile/{id}:
    get:
      tags:
        - Пользователи
      summary: получение пользователя по uuid
      security:
        - bearerAuth: [ ]
      description: "получение профиля пользовател по uuid. можно смотреть профили\
        \ только тех, с кем есть мэтч"
      parameters:
        - name: id
          in: path
          description: UUID пользователя
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                required:
                  - data
                  - success
                type: object
                properties:
                  success:
                    type: boolean
                    description: успех запроса
                    example: true
                  data:
                    $ref: '#/components/schemas/UserProfile'
        "403":
          description: Нет доступа для просмотра профиля
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: server error
          content: { }
  /profile/logout:
    post:
      tags:
        - Профиль пользователя
      summary: Выйти с устройства (удалить сессию)
      security:
        - bearerAuth: [ ]
      description: Удаление сессии и токенов для текущей сессии пользователя
      responses:
        "200":
          description: "OK, успешный выход"
          content:
            application/json:
              schema:
                required:
                  - success
                type: object
                properties:
                  success:
                    type: boolean
                    description: успех запроса
                    example: true
        "401":
          description: ошибка доступа
          content: { }
        "500":
          description: server error
          content: { }
  /profile/set-device-id:
    post:
      tags:
        - Профиль пользователя
      summary: Сохранить/обновить device ID
      security:
        - bearerAuth: [ ]
      description: добавляет или обновляет device id для текущей сессии авторизованного
        пользователя
      requestBody:
        content:
          application/json:
            schema:
              required:
                - deviceID
              type: object
              properties:
                deviceID:
                  type: string
                  description: device ID
                  example: abcd123456
        required: true
      responses:
        "200":
          description: "OK, device ID сохранён"
          content:
            application/json:
              schema:
                required:
                  - success
                type: object
                properties:
                  success:
                    type: boolean
                    description: успех запроса
                    example: true
        "500":
          description: server error
          content: { }
      x-codegen-request-body-name: data
...
components:
  schemas:
...
    ErrorResponse:
      required:
        - error
        - success
      type: object
      properties:
        success:
          type: boolean
          description: request success
          example: false
        error:
          required:
            - code
            - message
          type: object
          properties:
            code:
              type: integer
              description: internal error code
              example: 1001
            message:
              type: string
              description: error text
              example: example error
    UserProfile:
      required:
        - ID
      type: object
      properties:
        ID:
          type: string
          description: uuid пользователя
        name:
          type: string
          description: имя
        sex:
          type: string
          description: пол
        birthday:
          type: string
          description: дата рождения
        city:
          type: string
          description: город
        about:
          type: string
          description: о себе
        photos:
          type: array
          items:
            $ref: '#/components/schemas/UserPhoto'
        distance:
          type: integer
          description: расстояние до пользователя в километрах
          example: 4389
    UserPhoto:
      required:
        - ID
        - URL
        - isMain
      type: object
      properties:
        ID:
          type: string
          description: uuid фотографии
          example: 6a7485ca-2020-470d-b0b5-04f3ff19f12f
        URL:
          type: string
          description: URL адрес картинки
          example: https://example.example/example.jpg
        isMain:
          type: boolean
          description: флаг основное фото профиля
          example: false
        createdAt:
          type: string
          description: дата добавления
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT